refactor(core): restore media layer feature forwarding
