+++ packages/core/src/dom/media/hls-js/index.ts
@@ -1,5 +1,4 @@
 import HlsJs, { type HlsConfig as HlsJsConfig } from 'hls.js';
-import { type MediaPreloadType, type MediaStreamType, MediaStreamTypes } from '../../../core/media/types';
 import { HTMLVideoElementHost } from '../html-video-element-host';
 import { hlsJsErrors } from './errors';
 import { hlsJsLive } from './live';
@@ -49,30 +48,6 @@ export class HlsJsMedia extends HTMLVideoElementHost<HlsJs> {
 
   override set src(value: string) {
     this.#engine?.loadSource(value);
-  }
-
-  get preload(): MediaPreloadType {
-    return this.next?.preload ?? 'metadata';
-  }
-
-  set preload(value: MediaPreloadType) {
-    if (this.next) this.next.preload = value;
-  }
-
-  get streamType(): MediaStreamType {
-    return this.next?.streamType ?? MediaStreamTypes.UNKNOWN;
-  }
-
-  set streamType(value: MediaStreamType) {
-    if (this.next) this.next.streamType = value;
-  }
-
-  get liveEdgeStart() {
-    return this.next?.liveEdgeStart ?? Number.NaN;
-  }
-
-  get targetLiveWindow() {
-    return this.next?.targetLiveWindow ?? Number.NaN;
   }
 
   override get target() {
