+++ packages/core/src/dom/media/native-hls/tests/stream-type.test.ts
@@ -1,20 +1,10 @@
 import { describe, expect, it, vi } from 'vitest';
-import { type MediaStreamType, MediaStreamTypes } from '../../../../core/media/types';
+import { MediaStreamTypes } from '../../../../core/media/types';
 import { HTMLVideoElementHost } from '../../html-video-element-host';
 import { nativeHlsStreamType } from '../stream-type';
 
-class TestNativeHlsMedia extends HTMLVideoElementHost {
-  get streamType(): MediaStreamType {
-    return this.next?.streamType ?? MediaStreamTypes.UNKNOWN;
-  }
-
-  set streamType(value: MediaStreamType) {
-    if (this.next) this.next.streamType = value;
-  }
-}
-
 function setup(duration: number = Number.NaN) {
-  const host = new TestNativeHlsMedia();
+  const host = new HTMLVideoElementHost();
   const extension = nativeHlsStreamType();
   extension.install(host);
   const video = document.createElement('video');
@@ -29,7 +19,7 @@ function setDuration(video: HTMLVideoElement, duration: number) {
 
 describe('nativeHlsStreamType', () => {
   it('starts as UNKNOWN before any signal', () => {
-    const host = new TestNativeHlsMedia();
+    const host = new HTMLVideoElementHost();
     nativeHlsStreamType().install(host);
     expect(host.streamType).toBe(MediaStreamTypes.UNKNOWN);
   });
